requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
amqp==5.1.1
anyio==3.6.2
astroid==2.12.13
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.2.0
backcall==0.2.0
bcrypt==4.0.1
billiard==3.6.4.0
celery==5.2.7
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
decorator==5.1.1
dill==0.3.6
dnspython==2.2.1
ecdsa==0.18.0
email-validator==1.3.0
executing==1.2.0
fastapi==0.88.0
fastapi-utils==0.2.1
flower==1.2.0
frozenlist==1.3.3
greenlet==2.0.1
h11==0.14.0
humanize==4.4.0
idna==3.4
ipython==8.8.0
isort==5.11.4
jedi==0.18.2
jose==1.0.0
kombu==5.2.4
lazy-object-proxy==1.9.0
matplotlib-inline==0.1.6
mccabe==0.7.0
multidict==6.0.4
parso==0.8.3
passlib==1.7.4
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.6.2
prometheus-client==0.15.0
prompt-toolkit==3.0.36
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pydantic==1.10.4
Pygments==2.14.0
pylint==2.15.9
python-dotenv==0.21.0
python-jose==3.3.0
python-multipart==0.0.5
pytz==2022.7.1
redis==4.4.1
requests==2.28.2
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.46
stack-data==0.6.2
starlette==0.22.0
tomli==2.0.1
tomlkit==0.11.6
tornado==6.2
traitlets==5.8.1
typing_extensions==4.4.0
urllib3==1.26.14
uvicorn==0.20.0
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
yarl==1.8.2
